+++ packages/ui/src/components/chat/composer/ui/ComposerFooter.tsx
@@ -60,6 +60,7 @@ export interface ComposerFooterProps {
     dictationActive: boolean;
 
     onOpenSettings?: () => void;
+    onAttachmentMenuOpenChange?: (open: boolean) => void;
     onPickLocalFiles: () => void;
     onOpenIssuePicker: () => void;
     onOpenPrPicker: () => void;
@@ -102,6 +103,7 @@ export function ComposerFooter(props: ComposerFooterProps) {
         isPermissionAutoAcceptInteractive,
         dictationActive,
         onOpenSettings,
+        onAttachmentMenuOpenChange,
         onPickLocalFiles,
         onOpenIssuePicker,
         onOpenPrPicker,
@@ -162,6 +164,7 @@ export function ComposerFooter(props: ComposerFooterProps) {
                                 openIssuePicker={onOpenIssuePicker}
                                 openPrPicker={onOpenPrPicker}
                                 onOpenSettings={onOpenSettings}
+                                onMenuOpenChange={onAttachmentMenuOpenChange}
                                 onOpenMobileSheet={onOpenAttachSheet}
                             />
                             {!isPiKernel ? (
@@ -245,6 +248,7 @@ export function ComposerFooter(props: ComposerFooterProps) {
                             openIssuePicker={onOpenIssuePicker}
                             openPrPicker={onOpenPrPicker}
                             onOpenSettings={onOpenSettings}
+                            onMenuOpenChange={onAttachmentMenuOpenChange}
                         />
                         <FocusModeButton
                             footerIconButtonClass={footerIconButtonClass}